+++ contracts/libs/IexecLibCore_v5.sol
@@ -86,6 +86,8 @@ library IexecLibCore_v5 {
         uint16 botFirst; // Max: 65535
         uint16 botSize; // Max: 65535
         bytes3 shortTag; // Max: 0b111111111111111111111111 (0xFFFFFF)
+        // Offset 5
+        address sponsor;
     }
     /**
      * Tasks