+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNEXT
+- Sponsor match orders boost. (#67)
 - Migrate `claim` tests to hardhat. (#65, #66)
 - Upgrade Poco2 to solidity v0.8 . (#63)
 - Use common helpers in Poco Boost integration tests. (#62)